+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Change Log {#changes}
 
+### ? - ?
+
+##### Fixes :wrench:
+
+- Fixed a bug that prevented `UCesiumPrimitiveFeaturesBlueprintLibrary::GetPrimitiveFeatures` from retrieving the features of instanced meshes.
+
 ### v2.21.0 Preview for Unreal Engine 5.7 - 2025-11-17
 
 ##### Additions :tada: